% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{shuffle reads}
\usage{
shuffle(reads, k = 2, n = 2)
}
\arguments{
\item{reads}{An object of \link[Biostrings:XStringSet-class]{BStringSet}.}

\item{k}{the k-let size.}

\item{n}{the number of random sequences to generate.}
}
\value{
An object of \link[Biostrings:XStringSet-class]{BStringSet}.
}
\description{
Uses the uShuffle library to shuffle reads
}
\examples{
f <- system.file("extdata", "test.fa", package="uShuffleR")
shuffle(f)
}
